+++ cmd/main.go
@@ -70,8 +70,9 @@ func main() {
 	}
 
 	if err = (&controller.SleepScheduleReconciler{
-		Client: mgr.GetClient(),
-		Scheme: mgr.GetScheme(),
+		Client:   mgr.GetClient(),
+		Scheme:   mgr.GetScheme(),
+		Recorder: mgr.GetEventRecorder("slumlord-sleep-schedule"),
 	}).SetupWithManager(mgr); err != nil {
 		setupLog.Error(err, "unable to create controller", "controller", "SlumlordSleepSchedule")
 		os.Exit(1)